+++ vitest.d.ts
@@ -27,6 +27,5 @@ interface CustomMatchers<R = unknown, T = unknown> {
 }
 
 declare module 'vitest' {
-  interface Assertion<T = any> extends CustomMatchers<T> {}
-  interface AsymmetricMatchersContaining extends CustomMatchers {}
+  interface Matchers<R, T> extends CustomMatchers<R, T> {}
 }